Extract and share unnecessary effect diagnostic/refactor logic
