+++ src/diagnostics/unnecessaryEffectGen.ts
@@ -2,7 +2,7 @@ import * as Option from "effect/Option"
 import type ts from "typescript"
 import type { ApplicableDiagnosticDefinition } from "../definition.js"
 import { createDiagnostic } from "../definition.js"
-import * as TypeParser from "../utils/TypeParser.js"
+import * as AST from "../utils/AST.js"
 
 export const unnecessaryEffectGen = createDiagnostic({
   code: 5,
@@ -12,27 +12,8 @@ export const unnecessaryEffectGen = createDiagnostic({
     const brokenGenerators = new Set<ts.Node>()
 
     const visit = (node: ts.Node) => {
-      // did we hit an effect gen-like?
-      const effectGenLike = TypeParser.effectGen(ts, typeChecker)(node)
-
-      if (Option.isSome(effectGenLike)) {
-        // if we hit an effect gen-like, we need to check if its body is just a single return statement
-        const body = effectGenLike.value.body
-        if (
-          body.statements.length === 1 &&
-          ts.isReturnStatement(body.statements[0]) &&
-          body.statements[0].expression &&
-          ts.isYieldExpression(body.statements[0].expression) &&
-          body.statements[0].expression.expression
-        ) {
-          // get the type of the node
-          const nodeToCheck = body.statements[0].expression.expression
-          const type = typeChecker.getTypeAtLocation(nodeToCheck)
-          const maybeEffect = TypeParser.effectType(ts, typeChecker)(type, nodeToCheck)
-          if (Option.isSome(maybeEffect)) {
-            brokenGenerators.add(node)
-          }
-        }
+      if (Option.isSome(AST.getSingleReturnEffectFromEffectGen(typeChecker, node))) {
+        brokenGenerators.add(node)
       }
       ts.forEachChild(node, visit)
     }
